자료형 수정

+++ src/com/javaex/practice/Ex35.java
@@ -2,8 +2,8 @@ package com.javaex.practice;
 
 public class Ex35 {
 	public static void main(String[] args) {
-		final double speed = 300000;
-		double speed1 = (speed*365*24*60*60);
+		final long speed = 300000L;
+		long speed1 = (speed*365*24*60*60);
 		System.out.println("빛이 1년 동안 가는 거리: "+speed1+" km 입니다.");
 	}
 }
